PythonRpcServer/requirements.txt
backcall==0.1.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
decorator==4.4.2
ffmpy==0.2.2
grpcio==1.28.1
grpcio-tools==1.28.1
idna==2.9
ipython==7.13.0
ipython-genutils==0.2.0
jedi==0.17.0
KalturaApiClient==15.20.0
numpy==1.18.2
opencv-python==4.2.0.34
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.5
protobuf==3.11.3
ptyprocess==0.6.0
Pygments==2.6.1
pytube3==9.6.4
requests==2.23.0
requests-toolbelt==0.9.1
scenedetect==0.5.2
six==1.14.0
tqdm==4.45.0
traitlets==4.3.3
urllib3==1.25.8
wcwidth==0.1.9